Revert "feat: goodbye noir-starter, hello noirenberg"
