+++ package.json
@@ -1,23 +1,15 @@
 {
-  "name": "noirenberg",
+  "name": "create-noir",
   "version": "0.1.1",
   "type": "module",
-  "description": "This repo contains starter projects using Noir and Barretenberg, together with a CLI tool to help you get started.",
+  "description": "This is a reference repo to help you get started with writing zero-knowledge circuits with [Noir](https://noir-lang.org/).",
   "bin": "npx.js",
   "author": "",
   "license": "ISC",
-  "scripts": {
-    "start": "bunx tsx npx.ts",
-    "compile": "tsc npx.ts --esModuleInterop true --module nodenext && chmod +x npx.js",
-    "publish": "bun compile && npm publish"
-  },
   "dependencies": {
     "@inquirer/input": "^1.2.16",
     "@inquirer/select": "^1.3.3",
-    "axios": "^1.7.7",
     "commander": "^11.1.0",
-    "log-symbols": "^7.0.0",
-    "ora": "^8.1.0",
     "tiged": "^2.12.6"
   }
 }
